requirements-travis.txt
asgiref
certifi
cffi
chardet
cryptography
defusedxml
dj-database-url
django-allauth
django-crispy-forms
django-heroku
gunicorn
idna
oauthlib
psycopg2
pycparser
PyJWT
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
sqlparse
urllib3
whitenoise
